Pre-switch into posgres database

+++ client/src/app/members/photo-editor/photo-editor.component.ts
@@ -5,6 +5,7 @@ import { Member } from 'src/app/_models/member';
 import { Photo } from 'src/app/_models/photos';
 import { User } from 'src/app/_models/user';
 import { AccountService } from 'src/app/_services/account.service';
+import { ConfirmService } from 'src/app/_services/confirm.service';
 import { MembersService } from 'src/app/_services/members.service';
 import { environment } from 'src/environments/environment';
 
@@ -20,7 +21,7 @@ export class PhotoEditorComponent implements OnInit {
   baseUrl = environment.apiUrl;
   user: User;
 
-  constructor(private accountService: AccountService, private memberService: MembersService) { 
+  constructor(private accountService: AccountService, private memberService: MembersService, private confirmService: ConfirmService) { 
     this.accountService.currentUser$.pipe(take(1)).subscribe(user => this.user = user);
   }
 
@@ -45,9 +46,18 @@ export class PhotoEditorComponent implements OnInit {
   }
 
   deletePhoto(photoId: number) {
-    this.memberService.deletePhoto(photoId).subscribe(() => {
-      this.member.photos = this.member.photos.filter(x => x.id !== photoId);
+    this.confirmService.confirm('Confirm delete photo', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.memberService.deletePhoto(photoId).subscribe(() => {
+          this.member.photos = this.member.photos.filter(x => x.id !== photoId);
+        })
+      }
     })
+
+
+
+
+    
   }
 
   initializeUploader() {
